Require an `initialStep` to be defined
I wanted to come up with a way to avoid requiring this parameter but I
don't think it's possible. There's a bit of a chicken-and-egg problem
where

1. Render of the manager takes place
2. Rendering dictates the rendered steps
3. Rendered steps create define the initial step, if it's not provided
4. "Current step" gets passed back into each step

Which, ultimately, means that the `currentState` value changes twice in
a single render.  This is a no-go from Ember's perspective, causing
deprecated or unpredictable behavior.  So, for now at least, we have to
provide the initial step explicitly.