Transition to the first step when given `undefined`
If you're binding the current step to a query param, if you navigate
back to a URL without the query param, you'll end up passing `undefined`
into the state machine which will result in nothing happening, because
there are no steps called `undefined`.  Instead, actually want to
transition back to the first step.