Step manager now handles removal
When steps are destroyed, the willDestroy element hook now sends an action to remove a step. This removal does not happen immediately, instead it is placed in a queue to happen after render.
When steps are added, the addStep function is now modified to put steps to add in a queue as well, which will be acted upon during render
Then, when the render queue happens, the steps to add queue is flushed. First, we do not add steps that already exist. Then we proceed with the normal way to add steps. First, if there is no first step, it is set. Then, the current step is set if its not already. Finally, the last step added is always the last step.
For the flushRemovalQueue, its very similar. Steps that are to be added are not removed. If all the steps are to be removed, then we just reset the transitions and first/current/last steps. Otherwise we follow a similar procedure to removing steps. If the first/current/last step is removed, then new ones are found. That is it.