feat: expose array of step objects
Instead of exposing an array of `StepName` off of the yielded "wizard"
object, an array of `StepNode` is now provided. This allows users to
continue to access the `StepName` off of that object, but opens up the
ability to add additional API surface area to that object.

BREAKING CHANGE:

The exposed type of the `steps` array has changed, requiring a user to
reach into that object to extract the name if they need to.
Alternatively, they could use something like `ember-composable-helpers`
to get an array of `StepName` from the array of `StepNode`

```hbs
(map-by w.steps 'name')
```