feat: step context
A step can now be defined with a "context" that is also exposed in the array of
`StepNode` from the wizard. The "context" can be any value of any type;
whatever it is, it will be bound between the step where it is defined
and the step array where it is exposed.

Closes #104