refactor: remove TypeScript
It has been a constant battle against the lack of property type definitions
for Ember that is starting to place an artifical boundary on people contributing
to the project. While the TypeScript experiement was fun, I don't think it was
worth it overall.

BREAKING CHANGE: Package no longer provides TypeScript definitions for
  components. This is unlikely to _actually_ be a breaking change for
  anyone, but it is better to err on the safe side.