chore: remove activation hooks
At the time this feature was created, it made sense to support hooks that were called when a step was activated or deactivated. However, with the advent of more flexible element modifiers, this is no longer necessary as it is easy to re-create the same behavior using features that Ember provides official support for.

Closes #120 - Adding documentation for hooks is no longer necessary
Closes #129 - `did-insert` will not have the problem experiences on first render like `onActivate` did

BREAKING CHANGE: `onActivate` and `onDeactivate` should instead be called by `did-insert` or `will-destroy` on element inside the step definition.